- echobot and echoapp now compile without warnings.

+++ src/Core/include/K/Core/ReusableItems.h
@@ -141,7 +141,7 @@ ReusableItems<T>::ReusableItems(int numGroups)
 
 template<class T>
 T &ReusableItems<T>::Get(int groupId, int *itemId) {
-    assert((groupId >= 0) && (groupId < groupAnchors_.size()));
+    assert((groupId >= 0) && (groupId < static_cast<int>(groupAnchors_.size())));
     
     if (!idleCount_)
         AddIdleItem();
